fixed ugly code
