+++ src/raycaster/raycast.c
@@ -6,7 +6,7 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/17 12:34:34 by abied-ch          #+#    #+#             */
-/*   Updated: 2024/01/16 12:44:30 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/18 16:59:03 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
