+++ src/raycaster/texture_utils.c
@@ -6,7 +6,7 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/21 10:31:37 by abied-ch          #+#    #+#             */
-/*   Updated: 2024/01/18 16:01:11 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/18 16:59:24 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,13 +36,13 @@ void	draw_background(t_data *data)
 void	set_texture(t_data *data, t_txtr *texture)
 {
 	if (data->hit == NORTH)
-		*texture = data->wall1;
+		*texture = data->wall5;
 	else if (data->hit == EAST)
-		*texture = data->wall2;
+		*texture = data->wall6;
 	else if (data->hit == SOUTH)
-		*texture = data->wall3;
+		*texture = data->wall7;
 	else if (data->hit == WEST)
-		*texture = data->wall4;
+		*texture = data->wall8;
 }
 
 int	get_pixel(t_txtr *texture, int x, int y)
